fix(runtime): reconcile goal3 maintainability gates
