+++ contracts/functional-scenario-evidence.json
@@ -8,12 +8,6 @@
       ]
     },
     {
-      "test": "tests/functional/factory/visualization/runtime_metrics/end_to_end_costs_test.go::TestRuntimeCostsCachedInputConfigurationMatrix",
-      "stableIds": [
-        "cli/you.metrics.costs"
-      ]
-    },
-    {
       "test": "tests/functional/factory/visualization/runtime_metrics/end_to_end_costs_test.go::TestRuntimeCostsEndToEndFromProviderCompletion",
       "stableIds": [
         "cli/you.metrics.costs"
