+++ pkg/services/recordings/internal/events/event_history.go
@@ -309,49 +309,6 @@ func (h *FactoryEventHistory) AddEventTypeRecorder(recorder func(interfaces.Fact
 	h.mu.Unlock()
 }
 
-// AddDeferredSessionCompletionRecorder registers a callback for the
-// transport-facing terminal signal. Unlike AddEventTypeRecorder, this callback
-// is held until the owner explicitly publishes the already-appended
-// SESSION_COMPLETED after its durability boundary has succeeded.
-func (h *FactoryEventHistory) AddDeferredSessionCompletionRecorder(recorder func()) {
-	if h == nil || recorder == nil {
-		return
-	}
-	h.mu.Lock()
-	if h.deferredSessionCompletionPublished && h.hasSessionCompleted {
-		h.mu.Unlock()
-		recorder()
-		return
-	}
-	h.deferredSessionCompletionRecorders = append(h.deferredSessionCompletionRecorders, recorder)
-	if h.hasSessionCompleted {
-		h.deferredSessionCompletionPending = true
-	}
-	h.mu.Unlock()
-}
-
-// PublishDeferredSessionCompletion releases the terminal callbacks exactly
-// once. Callers invoke it only after the existing SESSION_COMPLETED event has
-// crossed the required durable recording boundary.
-func (h *FactoryEventHistory) PublishDeferredSessionCompletion() {
-	if h == nil {
-		return
-	}
-	h.mu.Lock()
-	if !h.deferredSessionCompletionPending || h.deferredSessionCompletionPublished {
-		h.mu.Unlock()
-		return
-	}
-	h.deferredSessionCompletionPublished = true
-	h.deferredSessionCompletionPending = false
-	recorders := append([]func(){}, h.deferredSessionCompletionRecorders...)
-	h.deferredSessionCompletionRecorders = nil
-	h.mu.Unlock()
-	for _, recorder := range recorders {
-		recorder()
-	}
-}
-
 // RecordInitialStructure records the static topology before work events.
 func (h *FactoryEventHistory) RecordInitialStructure() {
 	if h == nil {
