+++ tests/functional/internal/support/provider.go
@@ -90,24 +90,7 @@ func CodexSuccessStdout(result string) []byte {
 // CodexSuccessStdoutWithUsage emits the same sanitized Codex JSONL fixture as
 // CodexSuccessStdout while allowing a functional test to prove provider usage
 // metadata at the command-runner edge.
-func CodexSuccessStdoutWithUsage(result string, inputTokens, outputTokens int64) []byte {
-	return codexSuccessStdoutWithTokenUsage(result, map[string]any{
-		"input_tokens":  inputTokens,
-		"output_tokens": outputTokens,
-	})
-}
-
-// CodexSuccessStdoutWithUsageAndCachedInput emits a sanitized Codex JSONL
-// fixture with the cached-input subclass needed by runtime cost proofs.
-func CodexSuccessStdoutWithUsageAndCachedInput(result string, inputTokens, cachedInputTokens, outputTokens int64) []byte {
-	return codexSuccessStdoutWithTokenUsage(result, map[string]any{
-		"input_tokens":        inputTokens,
-		"cached_input_tokens": cachedInputTokens,
-		"output_tokens":       outputTokens,
-	})
-}
-
-func codexSuccessStdoutWithTokenUsage(result string, usage map[string]any) []byte {
+func CodexSuccessStdoutWithUsage(result string, inputTokens, outputTokens int64, cachedInputTokens ...int64) []byte {
 	if result == "" {
 		result = "Done. COMPLETE"
 	}
@@ -121,6 +104,13 @@ func codexSuccessStdoutWithTokenUsage(result string, usage map[string]any) []byt
 	})
 	if err != nil {
 		panic(err)
+	}
+	usage := map[string]any{
+		"input_tokens":  inputTokens,
+		"output_tokens": outputTokens,
+	}
+	if len(cachedInputTokens) > 0 {
+		usage["cached_input_tokens"] = cachedInputTokens[0]
 	}
 	turnCompleted, err := json.Marshal(map[string]any{
 		"type":  "turn.completed",
